requirements.txt
llama-index
pypdf
tqdm
chromadb
sentence-transformers
groq
python-dotenv
streamlit